Corrigiendo Los Deletes

+++ app/src/main/java/com/example/luisamaury/operativos_pia/DeleteUserActivity.java
@@ -11,12 +11,14 @@ import android.widget.EditText;
 import android.widget.Toast;
 
 import com.example.luisamaury.operativos_pia.R;
+import com.example.luisamaury.operativos_pia.materia.DeleteSubjectActivity;
 
 public class DeleteUserActivity extends AppCompatActivity {
     MyDBHandler myDb;
     Button btnDelete;
     EditText editDelete;
     String check;
+    boolean valido;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -60,12 +62,17 @@ public class DeleteUserActivity extends AppCompatActivity {
         return valid;
     }
     public void confirmDelete(){
-        Integer deletedRows = myDb.deleteUser(editDelete.getText().toString());
-        if (deletedRows > 0)
-            Toast.makeText(DeleteUserActivity.this, "Data Deleted", Toast.LENGTH_LONG).show();
-        else
-            Toast.makeText(DeleteUserActivity.this, "Data not Deleted", Toast.LENGTH_LONG).show();
-        editDelete.setText("");
+        valido = myDb.checkUserStudent(editDelete.getText().toString());
+        if(valido) {
+            Integer deletedRows = myDb.deleteUser(editDelete.getText().toString());
+            if (deletedRows > 0)
+                Toast.makeText(DeleteUserActivity.this, "Data Deleted", Toast.LENGTH_LONG).show();
+            else
+                Toast.makeText(DeleteUserActivity.this, "Data not Deleted", Toast.LENGTH_LONG).show();
+            editDelete.setText("");
+        }else{
+            Toast.makeText(DeleteUserActivity.this, "Cannot Delete, it is assigned to a student", Toast.LENGTH_LONG).show();
+        }
     }
 
 }
